requirements.txt
# Core Framework
PySide6>=6.6.0
SQLAlchemy>=2.0.0
Alembic>=1.13.0

# Database
# For SQLite (built-in, no install needed)
# For PostgreSQL (optional, for production):
# psycopg2-binary>=2.9.0

# Configuration & Utilities
PyYAML>=6.0
python-dateutil>=2.8.0

# PDF Generation
reportlab>=4.0.0

# Data Export
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel export support

# Development & Testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0  # Code formatter
# flake8>=6.0.0  # Linter
